streamlit/pyproject.toml
[tool.poetry]
name = "vizualizations"
version = "0.1.0"
description = ""
authors = ["Emma Berniczky <e.berniczky@gmail.com>"]
packages = [{ include = "pages" }]

[tool.poetry.dependencies]
python = "~3.12.8"
matplotlib = "^3.10.0"
plotly = "^5.24.1"
streamlit = "^1.41.1"
google-cloud-bigquery = ">=3.35.1,<4.0.0"
pandas = ">=2.3.1,<3.0.0"
db-dtypes = "^1.0.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
